requirements.txt
asgiref==3.11.0
Django==6.0
pillow==12.1.0
python-dotenv==1.2.1
sqlparse==0.5.5
pytest-django
pytest-cov
model_bakery    # Provides random data for tests